fix suite aux retours de Quentin : refacto
